+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the tool_vault plugin will be documented in this file.
 ### Added
 - Shutdown handler to detect if the process was cancelled
 - Capture all output and show in the vault log
+- Setting to force developer debugging during backup/restore
 ### Fixed
 - Fixed error when restoring data into tables with reserved words as fields if backup and
   restore sites are on different database engines (mysql/postgres)